rename qpu -> quam
